Move enumerations to features
