+++ src/features/enumerations/api/getEnumerations/contract.ts
@@ -1,5 +1,5 @@
-import { EnumerationsLookup } from '../../shared/types';
-import { RequestMetadata } from '../../shared/http';
+import { EnumerationsLookup } from 'shared/types';
+import { RequestMetadata } from 'shared/http';
 
 export const metadata: RequestMetadata = {
     url: '/api/enumerations',
