+++ src/features/enumerations/state/actions.ts
@@ -1,7 +1,7 @@
 import { Dispatch } from 'redux';
-import { getEnumerations } from '../../api/getEnumerations';
+import { getEnumerations } from '../api/getEnumerations';
 import { EnumerationsAction } from './types';
-import { AppState } from '..';
+import { AppState } from 'store';
 
 export const loadEnumerations = () =>
     async (dispatch: Dispatch<EnumerationsAction>, getState: () => AppState) => {
