+++ src/store/createStore.ts
@@ -1,14 +1,15 @@
 import { createStore, applyMiddleware, combineReducers } from 'redux';
-import { reducer as activityListReducer } from '../features/activityOverview/state/activityList/reducer';
-import { reducer as timingsPageReducer }  from '../features/activityOverview/state/timingList/reducer';
-import { reducer as timingsFormReducer }  from '../features/timeEntry/state/timing/reducer';
-import { reducer as issuesReducer } from 'features/timeEntry/state/issues/reducer';
-import { reducer as enumerationsReducer } from './enumerations/reducer';
-import { reducer as authReducer } from '../features/auth/state/reducer';
 import { reducer as globalReducer } from './@global/reducer';
 import { createThunkMiddleware } from './thunk';
 import { sumReducers } from './composition';
 import { errorHandler } from './errorHandler';
+
+import { reducer as activityListReducer } from 'features/activityOverview/state/activityList/reducer';
+import { reducer as timingsPageReducer }  from 'features/activityOverview/state/timingList/reducer';
+import { reducer as timingsFormReducer }  from 'features/timeEntry/state/timing/reducer';
+import { reducer as issuesReducer } from 'features/timeEntry/state/issues/reducer';
+import { reducer as enumerationsReducer } from 'features/enumerations/state/reducer';
+import { reducer as authReducer } from 'features/auth/state/reducer';
 
 
 export const store = createStore(
